requirements.txt
Flask==1.1.*
pytest==6.0.*
pytest-mock==3.3.*
pytest-xdist==1.26.*
pytest-cov==2.10.*
sqlalchemy==1.3.*
psycopg2-binary==2.8.6
vcrpy-unittest==0.1.7
flask-sqlalchemy==2.4.4
flask-migrate==2.5.3
python-dotenv==0.14.0
requests[security]==2.22.*
responses==0.12.0